1) MistralIntegration API update

+++ dataset/output/MISTRAL_codestral-latest/Task25_MISTRAL_codestral-latest.cpp
@@ -1,9 +1,9 @@
-#include <string>
 #include <vector>
+#include <string>
 
 std::vector<std::string> number(const std::vector<std::string>& lines) {
     std::vector<std::string> numberedLines;
-    for (size_t i = 0; i < lines.size(); i++) {
+    for (int i = 0; i < lines.size(); i++) {
         numberedLines.push_back(std::to_string(i + 1) + ": " + lines[i]);
     }
     return numberedLines;
